requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
django-extensions
django-hosts
django-fernet-fields
Pillow
djangorestframework
markdown
django-filter
six
django-cors-headers
python-decouple